requirements.txt
PyExecJS==1.5.1
requests==2.28.1